pyproject.toml
[tool.poetry]
name = "ambra-sdk"
version = "3.21.5.0-1"
description = "Ambrahealth python SDK"
authors = [
	"Ambrahealth AI team <python-sdk@ambrahealth.com>",
]
maintainers = [
	"Alexander Kapustin <akapustin@ambrahealth.com>",
]
readme = "README.md"
license = 'Apache-2.0'

# pypi url
homepage = 'https://github.com/dicomgrid/sdk-python'
repository = 'https://github.com/dicomgrid/sdk-python'
documentation = 'https://dicomgrid.github.io/sdk-python/index.html'

keywords = ['ambrahealth', 'ambra']

classifiers = [
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
bump_release_candidate = "release:bump_release_candidate"
start_release = "release:start_release"
bump_release = "release:bump_release"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.24.0"
aiohttp = "^3.6.2"
python-box = "^5.1.1"
pydicom = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
mypy = "^0.782"
pytest-cov = "^2.10.1"
wemake-python-styleguide = "^0.14.1"
dynaconf = "^3.1.0"
beautifulsoup4 = "^4.8.2"
jinja2 = "^2.11.1"
requests-mock = "^1.8.0"
sphinx = "^3.2.1"
tox = "^3.19.0"
tomlkit = "^0.5.11"
gitpython = "^3.1.0"
sphinxcontrib-spelling = "^5.3.0"
pytest-asyncio = "^0.15.1"
aioresponses = "^0.7.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
